created-common-util

+++ defi-backend-api/src/main/java/defi_backend_api/repository/UserDetailsRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserDetailsRepository extends JpaRepository<UserDetailsBO, Long> {
+public interface UserDetailsRepository extends JpaRepository<UserDetailsBO, Integer> {
 
     Optional<UserDetailsBO> findByEmail(String email);
 }
